requirements.txt
fastapi==0.111.1
pydantic-settings==2.3.4
sqlmodel==0.0.20
passlib==1.7.4
PyJWT==2.8.0
uvicorn==0.26.0
gunicorn==20.1.0
pymysql==1.1.1
emails